feat: 添加动态内存压缩训练脚本和评估基准
fix(qwen2_dymem): 移除训练时不必要的float转换

refactor(compressor): 优化分布式训练时的权重保存逻辑

feat(train): 支持本地jsonl数据文件并扩展文本列选择

feat: 新增NIAH风格SFT数据生成脚本

feat: 添加文档记忆微调数据生成脚本

feat: 实现RULER和LongBench评估基准

+++ src/dymem/transformer/qwen2_dymem/qwen2_dymem.py
@@ -212,7 +212,7 @@ class Qwen2AttentionWithMem(Qwen2Attention):
                 evicted_hidden_states = prenormed_hidden[:, sink_len:evict_end_index, :]
                 # with torch.amp.autocast(device_type="cuda",enabled=False):
                 mem_out = self.compressor(
-                    evicted_hidden_states.float() if self.training else evicted_hidden_states,
+                    evicted_hidden_states if self.training else evicted_hidden_states,
                     cache_params=past_key_value.mem_cache if past_key_value is not None else None,
                     cache_position=cache_position,
                 )  # [B, EVC_T, H]
